Create gameboard, game and player object factories
Set up the gameboard IIFE, to create the players and hold the
gameboard. Then the game IIFE to play the game, and the createPlayer
factory function to create instances of a player (we will only need 2).
It is also being considered to shift the creation of players from the
gameboard object to the game object.