Separate logic of concern
There was a mistake in the creation of the "CreatePlayer" factory,
which was simply a constructor. It was changed to a factory. Furt-
hermore, we create the "Player" object instances in the "game" ob-
ject instead now, which separates the logic better. It is more na-
tural to place such logic within the aformentioned object than the
"gameboard" object.