Add logic to get and switch player
Initial player in a tic tac toe game is always 'X', so that is set
as the default, and then calling the 'switchPlayers' method allows
to switch the current player. This is ideally for later when the
user will place a marker on the board and we need a way to switch.