Add placing marker logic
The logic is consistent in the sense that it works if you call the
methods on 'game', however there are a few issues such as, while a
marker does not override another, the turns gets switched nonthel-
ess, as it counts it as a turn used. This needs to be fixed.