Implement logic to check for row win
We are checking if the current play that the user has made is a w-
in. We don't bother to check the whole board, just based on what
move the user has currently made. In the case that it is true we
return the player that has won. Although, no implementations yet
on resetting the board - so that will have to be made later, after
vertical and diagonal checks are made.