Add win message & restart button
This allows the user to be notified that the game recognises that
it is over, and that they can restart if they wish to. Right now
the board resets immediately, but we want to attach a listener to
the button such that when it is clicked, only THEN it is reset.