Implement logic to remove message on restart button click
This removes the win message and restarts the game completely so
the user can play another round. Also fixed the problem where the
board reset when the win message appears. However, now need a way
to prevent the user from making any more wins when one of either X
or O has won.