Fix remove restart button error
There was an error that when the restart button is being removed,
the query selector that targeted it couldn't find it. On further
investigation, it was realised that the restart button is actually
part of the text content of the win message, so when we reset the
win message we end up deleting the restart button as well.